+++ apps/riscv-tui/src/ui/popup/info.rs
@@ -10,11 +10,12 @@ use ratatui::{
 
 pub fn render_popup(f: &mut Frame, emu: &EmuState) {
     let items = vec![
-        ListItem::new(format!(" Dram Base: {:#010x}", emu.mach_info.dram_base)),
-        ListItem::new(format!(" Dram Size: {} GB", emu.mach_info.dram_size)),
-        ListItem::new(format!(" Page Size: {} KB", emu.mach_info.page_size)),
-        ListItem::new(format!("  Hit Rate: {} %", emu.mach_info.hit_rate)),
-        ListItem::new(format!("Current PC: {:#010x}", emu.mach_snap.pc)),
+        ListItem::new(format!(" Dram Base: {:#010x}", emu.mach_snap.info.dram_base)),
+        ListItem::new(format!(" Dram Size: {} GB", emu.mach_snap.info.dram_size)),
+        ListItem::new(format!(" Page Size: {} KB", emu.mach_snap.info.page_size)),
+        ListItem::new(format!(" TLB Hit Rate: {} %", emu.mach_snap.info.hit_rate)),
+        ListItem::new(format!(" Current Mode: {}", emu.mach_snap.info.curr_mode)),
+        ListItem::new(format!(" Current PC: {:#010x}", emu.mach_snap.pc)),
     ];
 
     let list = List::new(items)
@@ -27,8 +28,8 @@ pub fn render_popup(f: &mut Frame, emu: &EmuState) {
 }
 
 fn popup_area(area: Rect) -> Rect {
-    let vertical = Layout::vertical([Constraint::Length(7)]).flex(Flex::End);
-    let horizontal = Layout::horizontal([Constraint::Length(25)]).flex(Flex::End);
+    let vertical = Layout::vertical([Constraint::Length(8)]).flex(Flex::End);
+    let horizontal = Layout::horizontal([Constraint::Length(30)]).flex(Flex::End);
     let [area] = vertical.areas(area);
     let [area] = horizontal.areas(area);
     area